+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+- Added `num_rows`, `duration_ms`, and `command` attributes to `Lotus.Result` struct returned by query execution
+
 ## [0.8.0] - 2025-09-03
 
 ### Added